core/middleware.py
"""
Multi-tenant middleware for Mizan AI
Ensures tenant isolation and context validation for every request
"""
from django.http import JsonResponse
from django.utils.deprecation import MiddlewareMixin
from rest_framework_simplejwt.authentication import JWTAuthentication
from rest_framework.exceptions import AuthenticationFailed
import logging

logger = logging.getLogger(__name__)


class TenantContextMiddleware(MiddlewareMixin):
    """
    Middleware that validates and injects tenant context into every request.
    
    Flow:
    1. Extract JWT token from request
    2. Validate user authentication
    3. Extract tenant_id (restaurant) from user
    4. Inject tenant context into request
    5. Validate user has access to requested tenant resources
    """
    
    def process_request(self, request):
        # Skip for public endpoints
        public_paths = [
            '/api/auth/login',
            '/api/auth/register',
            '/api/auth/refresh',
            '/api/invitations/accept',
            '/admin/',
            '/api/docs/',
            '/api/schema/',
        ]
        
        if any(request.path.startswith(path) for path in public_paths):
            return None
        
        # Skip for non-API requests
        if not request.path.startswith('/api/'):
            return None
        
        # Extract and validate JWT token
        jwt_auth = JWTAuthentication()
        try:
            validated_token = jwt_auth.get_validated_token(
                jwt_auth.get_raw_token(jwt_auth.get_header(request))
            )
            user = jwt_auth.get_user(validated_token)
            
            # Inject tenant context
            if hasattr(user, 'restaurant') and user.restaurant:
                request.tenant_id = str(user.restaurant.id)
                request.tenant = user.restaurant
                request.tenant_name = user.restaurant.name
            else:
                # User has no restaurant association
                if user.is_superuser:
                    # Superusers can access without tenant
                    request.tenant_id = None
                    request.tenant = None
                    request.tenant_name = 'System Admin'
                else:
                    logger.warning(f"User {user.email} has no restaurant association")
                    return JsonResponse({
                        'error': 'No restaurant association',
                        'detail': 'User must be associated with a restaurant'
                    }, status=403)
            
            # Log tenant context for debugging
            logger.debug(f"Tenant context: {request.tenant_name} ({request.tenant_id})")
            
        except (AuthenticationFailed, AttributeError, TypeError) as e:
            # Authentication failed or no token provided
            logger.debug(f"Authentication failed: {str(e)}")
            # Let the view handle authentication
            pass
        
        return None
    
    def process_view(self, request, view_func, view_args, view_kwargs):
        """
        Validate tenant access for specific resources
        """
        # If tenant context exists, validate access
        if hasattr(request, 'tenant_id') and request.tenant_id:
            # Check if view is accessing tenant-specific resource
            # This is handled by view permissions, but we can add extra validation here
            pass
        
        return None


class TenantIsolationMiddleware(MiddlewareMixin):
    """
    Ensures data isolation between tenants at the database query level.
    This is a safety net in addition to view-level filtering.
    """
    
    def process_request(self, request):
        # Store original tenant context in thread-local storage if needed
        # This can be used by model managers to automatically filter queries
        if hasattr(request, 'tenant_id'):
            # Set thread-local tenant context
            from threading import local
            _thread_locals = local()
            _thread_locals.tenant_id = request.tenant_id
        
        return None


class AuditLoggingMiddleware(MiddlewareMixin):
    """
    Logs all tenant actions for audit trail
    """
    
    def process_response(self, request, response):
        # Log tenant actions for audit
        if hasattr(request, 'tenant_id') and hasattr(request, 'user'):
            if request.method in ['POST', 'PUT', 'PATCH', 'DELETE']:
                logger.info(
                    f"Tenant Action: {request.method} {request.path} | "
                    f"Tenant: {request.tenant_name} | "
                    f"User: {request.user.email} | "
                    f"Status: {response.status_code}"
                )
        
        return response