requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
django-filter==23.3
psycopg2-binary==2.9.9
python-decouple==3.8
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
firebase-admin==6.2.0
pillow==10.1.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.21.0
drf-spectacular==0.26.5

# AI/ML Libraries
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0

# Additional utilities
python-dateutil==2.8.2